+++ test/support/paradex_app/postgrex_types.ex
@@ -1,5 +1,5 @@
 Postgrex.Types.define(
   ParadexApp.PostgrexTypes,
-  Pgvector.extensions() ++ Ecto.Adapters.Postgres.extensions(),
+  Pgvector.extensions() ++ Paradex.extensions() ++ Ecto.Adapters.Postgres.extensions(),
   []
 )